// UnsDashboard/unsdashboard.client/src/opcua/models/PubSubConnectionDataType.ts
export interface PubSubConnectionDataType {
    /**
     * 
     * @type {string}
     * @memberof PubSubConnectionDataType
     */
    Name?: string;
    /**
     * 
     * @type {boolean}
     * @memberof PubSubConnectionDataType
     */
    Enabled?: boolean;
    /**
     * 
     * @type {object}
     * @memberof PubSubConnectionDataType
     */
    PublisherId?: object;
    /**
     * 
     * @type {string}
     * @memberof PubSubConnectionDataType
     */
    TransportProfileUri?: string;
    /**
     * 
     * @type {ExtensionObject}
     * @memberof PubSubConnectionDataType
     */
    Address?: ExtensionObject;
    /**
     * 
     * @type {Array<KeyValuePair>}
     * @memberof PubSubConnectionDataType
     */
    ConnectionProperties?: Array<KeyValuePair>;
    /**
     * 
     * @type {ExtensionObject}
     * @memberof PubSubConnectionDataType
     */
    TransportSettings?: ExtensionObject;
    /**
     * 
     * @type {Array<WriterGroupDataType>}
     * @memberof PubSubConnectionDataType
     */
    WriterGroups?: Array<WriterGroupDataType>;
    /**
     * 
     * @type {Array<ReaderGroupDataType>}
     * @memberof PubSubConnectionDataType
     */
    ReaderGroups?: Array<ReaderGroupDataType>;
}