// UnsDashboard/unsdashboard.client/src/opcua/models/DataSetMetaDataType.ts
export interface DataSetMetaDataType {
    /**
     * 
     * @type {string}
     * @memberof DataSetMetaDataType
     */
    Name?: string;
    /**
     * 
     * @type {LocalizedText}
     * @memberof DataSetMetaDataType
     */
    Description?: LocalizedText;
    /**
     * 
     * @type {Array<FieldMetaData>}
     * @memberof DataSetMetaDataType
     */
    Fields?: Array<FieldMetaData>;
    /**
     * 
     * @type {string}
     * @memberof DataSetMetaDataType
     */
    DataSetClassId?: string;
    /**
     * 
     * @type {ConfigurationVersionDataType}
     * @memberof DataSetMetaDataType
     */
    ConfigurationVersion?: ConfigurationVersionDataType;
    /**
     * 
     * @type {Array<string>}
     * @memberof DataSetMetaDataType
     */
    Namespaces?: Array<string>;
    /**
     * 
     * @type {Array<StructureDescription>}
     * @memberof DataSetMetaDataType
     */
    StructureDataTypes?: Array<StructureDescription>;
    /**
     * 
     * @type {Array<EnumDescription>}
     * @memberof DataSetMetaDataType
     */
    EnumDataTypes?: Array<EnumDescription>;
    /**
     * 
     * @type {Array<SimpleTypeDescription>}
     * @memberof DataSetMetaDataType
     */
    SimpleDataTypes?: Array<SimpleTypeDescription>;
}