// UnsDashboard/unsdashboard.client/src/opcua/models/WriterGroupDataType.ts
export interface WriterGroupDataType {
    /**
     * 
     * @type {number}
     * @memberof WriterGroupDataType
     */
    WriterGroupId?: number;
    /**
     * 
     * @type {number}
     * @memberof WriterGroupDataType
     */
    PublishingInterval?: number;
    /**
     * 
     * @type {number}
     * @memberof WriterGroupDataType
     */
    KeepAliveTime?: number;
    /**
     * 
     * @type {number}
     * @memberof WriterGroupDataType
     */
    Priority?: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof WriterGroupDataType
     */
    LocaleIds?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof WriterGroupDataType
     */
    HeaderLayoutUri?: string;
    /**
     * 
     * @type {ExtensionObject}
     * @memberof WriterGroupDataType
     */
    TransportSettings?: ExtensionObject;
    /**
     * 
     * @type {ExtensionObject}
     * @memberof WriterGroupDataType
     */
    MessageSettings?: ExtensionObject;
    /**
     * 
     * @type {Array<DataSetWriterDataType>}
     * @memberof WriterGroupDataType
     */
    DataSetWriters?: Array<DataSetWriterDataType>;
    /**
     * 
     * @type {string}
     * @memberof WriterGroupDataType
     */
    Name?: string;
    /**
     * 
     * @type {boolean}
     * @memberof WriterGroupDataType
     */
    Enabled?: boolean;
    /**
     * 
     * @type {number}
     * @memberof WriterGroupDataType
     */
    SecurityMode?: number;
    /**
     * 
     * @type {string}
     * @memberof WriterGroupDataType
     */
    SecurityGroupId?: string;
    /**
     * 
     * @type {Array<EndpointDescription>}
     * @memberof WriterGroupDataType
     */
    SecurityKeyServices?: Array<EndpointDescription>;
    /**
     * 
     * @type {number}
     * @memberof WriterGroupDataType
     */
    MaxNetworkMessageSize?: number;
    /**
     * 
     * @type {Array<KeyValuePair>}
     * @memberof WriterGroupDataType
     */
    GroupProperties?: Array<KeyValuePair>;
}