// UnsDashboard/unsdashboard.client/src/opcua/models/BrokerDataSetWriterTransportDataType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OPC UA REST API
 * This API provides simple REST based access to an OPC UA server.
 *
 * The version of the OpenAPI document: 0.0.1
 * Contact: office@opcfoundation.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BrokerDataSetWriterTransportDataType
 */
export interface BrokerDataSetWriterTransportDataType {
    /**
     * 
     * @type {string}
     * @memberof BrokerDataSetWriterTransportDataType
     */
    QueueName?: string;
    /**
     * 
     * @type {string}
     * @memberof BrokerDataSetWriterTransportDataType
     */
    ResourceUri?: string;
    /**
     * 
     * @type {string}
     * @memberof BrokerDataSetWriterTransportDataType
     */
    AuthenticationProfileUri?: string;
    /**
     * 
     * @type {number}
     * @memberof BrokerDataSetWriterTransportDataType
     */
    RequestedDeliveryGuarantee?: number;
    /**
     * 
     * @type {string}
     * @memberof BrokerDataSetWriterTransportDataType
     */
    MetaDataQueueName?: string;
    /**
     * 
     * @type {number}
     * @memberof BrokerDataSetWriterTransportDataType
     */
    MetaDataUpdateTime?: number;
}

/**
 * Check if a given object implements the BrokerDataSetWriterTransportDataType interface.
 */
export function instanceOfBrokerDataSetWriterTransportDataType(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function BrokerDataSetWriterTransportDataTypeFromJSON(json: any): BrokerDataSetWriterTransportDataType {
    return BrokerDataSetWriterTransportDataTypeFromJSONTyped(json, false);
}

export function BrokerDataSetWriterTransportDataTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): BrokerDataSetWriterTransportDataType {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'QueueName': !exists(json, 'QueueName') ? undefined : json['QueueName'],
        'ResourceUri': !exists(json, 'ResourceUri') ? undefined : json['ResourceUri'],
        'AuthenticationProfileUri': !exists(json, 'AuthenticationProfileUri') ? undefined : json['AuthenticationProfileUri'],
        'RequestedDeliveryGuarantee': !exists(json, 'RequestedDeliveryGuarantee') ? undefined : json['RequestedDeliveryGuarantee'],
        'MetaDataQueueName': !exists(json, 'MetaDataQueueName') ? undefined : json['MetaDataQueueName'],
        'MetaDataUpdateTime': !exists(json, 'MetaDataUpdateTime') ? undefined : json['MetaDataUpdateTime'],
    };
}

export function BrokerDataSetWriterTransportDataTypeToJSON(value?: BrokerDataSetWriterTransportDataType | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'QueueName': value.QueueName,
        'ResourceUri': value.ResourceUri,
        'AuthenticationProfileUri': value.AuthenticationProfileUri,
        'RequestedDeliveryGuarantee': value.RequestedDeliveryGuarantee,
        'MetaDataQueueName': value.MetaDataQueueName,
        'MetaDataUpdateTime': value.MetaDataUpdateTime,
    };
}